update vite server
