+++ app/routes/_dashboard/department/_index.tsx
@@ -19,7 +19,7 @@ import { useLoaderData } from '@remix-run/react'
 import UploadFormDialog from '../departments_.$id.details/components/form/upload-member-form'
 import { EditMemberForm } from '../departments_.$id.details/components/form/edit-member-form'
 import AttendanceFormDialog from '../../../shared/attendance-form/form/attendance-form'
-import { AttendanceReportEntity } from '@prisma/client'
+import { AttendanceReportEntity } from '~/shared/enum'
 import { Header } from '~/components/layout/header'
 import { FilterForm } from '~/shared/forms/filter-form'
 import { renderTable } from '~/shared/member-table/table.utlis'
