+++ app/routes/_dashboard/honor-family/_index.tsx
@@ -11,7 +11,7 @@ import { useHonorFamily } from './hooks/use-honor-family'
 import { useDownloadFile } from '~/shared/hooks'
 import { FilterFormDialog } from './components/forms/filter-form'
 import AttendanceFormDialog from '../../../shared/attendance-form/form/attendance-form'
-import { AttendanceReportEntity } from '@prisma/client'
+import { AttendanceReportEntity } from '~/shared/enum'
 import { EditMemberForm } from './components/forms/edit-member-form'
 import { UploadMemberForm } from './components/forms/upload-member-form'
 import { actionFn } from './server/action.server'
