+++ app/routes/_dashboard/tribe/_index.tsx
@@ -14,7 +14,7 @@ import { speedDialItems } from './constants'
 import { EditEntityMemberForm } from '~/shared/forms/edit-entity-member-form'
 import { actionFn } from './action.server'
 import { UploadFormDialog } from '~/shared/forms/upload-member-form'
-import { AttendanceReportEntity } from '@prisma/client'
+import { AttendanceReportEntity } from '~/shared/enum'
 import AttendanceFormDialog from '../../../shared/attendance-form/form/attendance-form'
 import { renderTable } from '~/shared/member-table/table.utlis'
 import { GeneralErrorBoundary } from '~/components/error-boundary'
