+++ package.json
@@ -2,15 +2,12 @@
   "name": "jeriel",
   "private": true,
   "sideEffects": false,
-  "packageManager": "pnpm@10.28.0",
+  "type": "module",
+  "packageManager": "pnpm@11.1.0",
   "scripts": {
-    "build": "run-s 'build:*'",
-    "build:remix": "remix vite:build",
-    "build:server": "esbuild --platform=node --format=cjs ./server.ts --outdir=build --bundle --external:fsevents --external:vite",
-    "dev": "run-p 'dev:*'",
-    "dev:server": "cross-env NODE_ENV=development npm run build:server -- --watch",
-    "dev:remix": "cross-env NODE_ENV=development node --watch-path=./build/server.js ./build/server.js",
-    "start": "cross-env NODE_ENV=production node ./build/server.js",
+    "build": "remix vite:build",
+    "dev": "remix vite:dev",
+    "start": "cross-env NODE_ENV=production node ./server.js",
     "format": "prettier --write .",
     "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
     "typecheck": "tsc --noEmit",
